csharp recodegen

+++ hosting/csharp/core/Models/CroppingSuggestionItem.cs
@@ -22,7 +22,7 @@ namespace Aliyun.SDK.Hosting.Client.Models
 
         [NameInMap("score")]
         [Validation(Required=false)]
-        public int? Score { get; set; }
+        public double? Score { get; set; }
 
     }
